refactor(organisation): rename Customer resource to Organisation
