--- nodes/Mittwald/resources/implementations/Customer/resource.ts
@@ -1,5 +1,0 @@
-import { Resource } from '../../base';
-
-export const customerResource = new Resource({
-	name: 'Customer',
-});
