+++ nodes/Mittwald/resources/implementations/Organisation/resource.ts
@@ -1,0 +1,5 @@
+import { Resource } from '../../base';
+
+export const organisationResource = new Resource({
+	name: 'Organisation',
+});
