+++ nodes/Mittwald/resources/implementations/operations.ts
@@ -3,6 +3,6 @@ import './App/operations';
 import './Server/operations';
 import './Conversation/operations';
 import './Contributor/operations';
-import './Customer/operations';
+import './Organisation/operations';
 import './Domain/operations';
 import './Database/operations';
